Move map's redux functions to redux folder
