+++ src/redux/reducers/index.js
@@ -1,5 +1,9 @@
 import unit from './unit';
 import filter from './filter';
+import locale from './locale';
+import map from './map';
 
 
-export default { unit, filter };
+export default {
+  unit, filter, locale, map,
+};
