+++ src/redux/reducers/map.js
