+++ src/redux/selectors/map.js
